Refactor crypto components with kebab-case filenames and cleanup
