+++ components/crypto/loading-sparkline.tsx
